fix: move PropertySelector
